made terrain able to modify player movement speed and push player

+++ Client/Tonpose/core/src/cs309/tonpose/Terrain.java
@@ -28,7 +28,7 @@ public class Terrain extends Entity {
         return modY;
     }
 
-    public double getScale(){
+    public float getScale(){
         return scale;
     }
 }
